:construction_worker: pre-build data into static file
