+++ src/components/main.jsx
@@ -1,10 +1,10 @@
 /* global window */
 /** @jsxImportSource @emotion/react */
-import {materials} from 'genshin-db';
 import {useCallback, useEffect, useState} from 'react';
 import {Global, css} from '@emotion/react';
 
 import storage, {fromLocalStorage} from '../utils/local-storage.js';
+import materials from '../data.json';
 import FarmHelper from './farm-helper.jsx';
 import ItemCategories from './item-categories.jsx';
 
@@ -58,42 +58,20 @@ const localStorageKey = 'genshin-farming-helper';
 let didRun = false;
 
 export default function Main() {
-	const defaultOptions = {matchCategories: true, verboseCategories: true};
-	const sundayDrops = materials('Sunday', defaultOptions).sort((a, b) => a.sortorder - b.sortorder);
-
-	const talentMaterials = sundayDrops.filter(material => material?.materialtype.startsWith('Talent'));
-	const talentMaterialsRare = talentMaterials.filter(material => Number.parseInt(material?.rarity, 10) > 3);
-
-	const weaponMaterials = sundayDrops.filter(material => material?.materialtype.startsWith('Weapon'));
-	const weaponMaterialsRare = weaponMaterials.filter(material => Number.parseInt(material?.rarity, 10) > 4);
-
-	const characterMaterials = materials('names', defaultOptions).filter(material => material?.materialtype.startsWith('Character Level-Up Material') || material?.materialtype.startsWith('Local')).sort((a, b) => a.sortorder - b.sortorder);
-
-	const characterAscensionMaterials = characterMaterials.filter(material => material?.description.startsWith('Character Ascension'));
-	const characterAscensionMaterialsRare = characterAscensionMaterials.filter(material => Number.parseInt(material?.rarity, 10) > 4);
-
-	const characterLVLMaterials = characterMaterials.filter(material => !material?.description.startsWith('Character Ascension'));
-	const characterLVLMaterialsRare = characterLVLMaterials.filter(material => {
+	const talentMaterialsRare = materials.talentMaterials.filter(material => Number.parseInt(material?.rarity, 10) > 3);
+	const weaponMaterialsRare = materials.weaponMaterials.filter(material => Number.parseInt(material?.rarity, 10) > 4);
+	const characterAscensionMaterialsRare = materials.characterAscensionMaterials.filter(material => Number.parseInt(material?.rarity, 10) > 4);
+	const characterLVLMaterialsRare = materials.characterLVLMaterials.filter(material => {
 		const rarityInt = Number.parseInt(material?.rarity, 10);
 		return rarityInt > 2 ? (rarityInt === 3 ? !material.source.includes('Stardust Exchange') : true) : false;
 	});
-
-	const localSpecialties = characterMaterials.filter(material => material?.materialtype.startsWith('Local'));
-
-	const materialsAll = {
-		talentMaterials,
-		weaponMaterials,
-		characterAscensionMaterials,
-		characterLVLMaterials,
-		localSpecialties,
-	};
 
 	const materialsRare = {
 		talentMaterials: talentMaterialsRare,
 		weaponMaterials: weaponMaterialsRare,
 		characterAscensionMaterials: characterAscensionMaterialsRare,
 		characterLVLMaterials: characterLVLMaterialsRare,
-		localSpecialties,
+		localSpecialties: materials.localSpecialties,
 	};
 
 	const [farmHelperList, setFarmHelperList] = useState([]);
@@ -118,12 +96,11 @@ export default function Main() {
 						key={item}
 						category={category}
 						item={item}
-						materials={materialsAll}
+						materials={materials}
 						onRemove={onRemove}
 					/>,
 				],
 		);
-		// eslint-disable-next-line react-hooks/exhaustive-deps
 	}, []);
 
 	const onChange = event => {
